LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/iframe-in-container-invalidation-expected.txt



FAIL @container-dependent elements respond to size changes of an @container-dependent iframe assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
